pl - rename custom-checkout.d.ts -> checkout.d.ts
