+++ types/stripe-js/checkout.d.ts
